Restructure rate component
