--- src/app/assign-rate/assign-rate-main.component.html
@@ -1,3 +1,0 @@
-<div class="animated fadeInUp row" style="animation-delay: 0.1s">
-
-</div>
