--- src/app/assign-rate/assign-rate-main.component.ts
@@ -1,13 +1,0 @@
-import {Component, OnInit} from '@angular/core';
-
-@Component({
-    selector: 'app-assign-rate-main',
-    templateUrl: './assign-rate-main.component.html'
-})
-
-export class AssignRateMainComponent implements OnInit {
-
-    ngOnInit(): void {
-    }
-
-}
