--- src/app/assign-rate/assign-rate.module.ts
@@ -1,16 +1,0 @@
-import {NgModule} from '@angular/core';
-import {AssignRateRoutes} from './assign-rate.routes';
-import {AssignRateMainComponent} from './assign-rate-main.component';
-
-@NgModule({
-    imports: [
-        AssignRateRoutes
-    ],
-    declarations: [
-        AssignRateMainComponent
-    ]
-})
-
-export class AssignRateModule {
-
-}
