--- src/app/assign-rate/assign-rate.routes.ts
@@ -1,9 +1,0 @@
-import {RouterModule} from '@angular/router';
-import {AssignRateMainComponent} from './assign-rate-main.component';
-
-const routes = [{
-    path : '',
-    component: AssignRateMainComponent
-}];
-
-export const AssignRateRoutes = RouterModule.forChild(routes);
