+++ src/app/rate/create/rate-category/category.component.scss
