+++ src/app/rate/create/rate-currency/currency.component.scss
