+++ src/app/rate/create/rate-main/rate-main.component.scss
