+++ src/app/rate/create/rate-ratecategory/ratecategory.component.html
@@ -59,7 +59,7 @@
 
                             <div class="form-group">
                                 <div class="col-sm-6 col-sm-offset-3">
-                                    <button type="submit" class="btn btn-primary animated fadeIn">Add</button>
+                                    <button type="submit" class="btn btn-primary animated fadeIn">Add Combination</button>
                                     <a class="btn btn-danger" (click)="clearCatSelection()">Clear</a>
                                     <div *ngIf="isMappigError">
                                         <p class="error">Invalid Mapping</p>
