+++ src/app/rate/create/rate-ratecategory/ratecategory.component.scss
@@ -1,8 +1,8 @@
-@import "/../../../styles/colors";
-@import "/../../../styles/mixing";
+@import "/../../../../styles/colors";
+@import "/../../../../styles/mixing";
 
 .subcategoryContainer {
-  background-color: #f1f3f5;
+  background-color: #cccccc;
   padding: 5px;
   font-size: 15px;
   margin: 10px;
