+++ src/app/rate/create/rate-tariff/tariff.component.scss
