+++ src/app/rate/rate.module.ts
@@ -1,17 +1,19 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {RateCategoryComponent} from './rate-ratecategory/ratecategory.component';
-import {CategoryComponent} from './rate-category/category.component';
-import {DefinitionComponent} from './rate-definition/definition.component';
-import {TariffComponent} from './rate-tariff/tariff.component';
-import {RateMainComponent} from './rate-main/rate-main.component';
+import {RateCategoryComponent} from './create/rate-ratecategory/ratecategory.component';
+import {CategoryComponent} from './create/rate-category/category.component';
+import {DefinitionComponent} from './create/rate-definition/definition.component';
+import {TariffComponent} from './create/rate-tariff/tariff.component';
+import {RateMainComponent} from './create/rate-main/rate-main.component';
+import {AssignRateMainComponent} from './assign/assign-rate-main.component';
 import {RateRoutes} from './rate.routes';
 import {SharedModule} from '../shared/shared.module';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import { ModalModule } from 'ngx-bootstrap/modal';
 import {TooltipModule} from 'ngx-bootstrap/tooltip';
 import { TypeaheadModule } from 'ngx-bootstrap/typeahead';
-import {CurrencyComponent} from './rate-currency/currency.component';
+import {CurrencyComponent} from './create/rate-currency/currency.component';
+import { DualListBoxModule } from 'ng2-dual-list-box';
 
 @NgModule({
     imports: [
@@ -22,7 +24,8 @@ import {CurrencyComponent} from './rate-currency/currency.component';
         ReactiveFormsModule,
         ModalModule.forRoot(),
         TypeaheadModule.forRoot(),
-        TooltipModule.forRoot()
+        TooltipModule.forRoot(),
+        DualListBoxModule.forRoot()
     ],
     declarations: [
         RateMainComponent,
@@ -30,7 +33,8 @@ import {CurrencyComponent} from './rate-currency/currency.component';
         CategoryComponent,
         CurrencyComponent,
         TariffComponent,
-        DefinitionComponent
+        DefinitionComponent,
+        AssignRateMainComponent
     ]
 })
 export class RateModule {
