+++ src/app/rate/rate.routes.ts
@@ -1,9 +1,16 @@
 import {RouterModule} from '@angular/router';
-import {RateMainComponent} from './rate-main/rate-main.component';
+import {RateMainComponent} from './create/rate-main/rate-main.component';
+import {AssignRateMainComponent} from './assign/assign-rate-main.component';
 
-const routes = [{
-  path : '',
-  component : RateMainComponent
-}];
+const routes = [
+    {
+        path: 'create',
+        component: RateMainComponent
+    },
+    {
+        path: 'assign',
+        component: AssignRateMainComponent
+    },
+];
 
 export const RateRoutes = RouterModule.forChild(routes);
